Player class separated
