+++ index.html
@@ -8,7 +8,7 @@
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css" integrity="sha384-rbsA2VBKQhggwzxH7pPCaAqO46MgnOM80zW1RWuH61DGLwZJEdK2Kadq2F9CUG65" crossorigin="anonymous">
     
     <link rel="stylesheet" href="padel-tournament-styles.css">
-    <script src="padel-tournament-script.js"></script>
+    <script type="module" src="padel-tournament-script.js"></script>
 
 </head>
